Added non-functional carachter flipping logic
Flipping collision shapes is not supported, so the sub components must be flipped. However Godot skeletons IK's break when flipped. So character flipping is not possible right now without creating a separate player skeleton and linked components for each direction.

I do not know if this is intended behavior or a bug, but it is documented here:
https://github.com/godotengine/godot/issues/75224
https://github.com/godotengine/godot/issues/80252