Finished first attempt at mining particle system - Unusable
Godot resets the particle system any time the 'amount' value is changed. It is unclear if this is expected behavior, but it is a long standing fact.
See : https://github.com/godotengine/godot/issues/16352
This implementation can not work in light of this. Will re-design for less efficient multi-node implementation, committing this in case it can be re-used in the future.