Save and reload player's last position.
This only saves if you properly exit using Esc, if you "crash out" of the game with the Godot Engine's stop button the position is not saved, which is fine, just so you know.

The default location is just 0, 0 which, along with the "find a safe spot" code should put you on solid ground in the center of the world. This may need raised or lowered if the world changes.

Rotation is not considered yet.

The "find a safe spot" code still needs testing and improving, so it could still do weird things.